+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3ClassTest.java
@@ -748,4 +748,25 @@ public class ActionScript3ClassTest extends ActionScript3DecompileTestBase {
                 + "}\n"
         );
     }
+
+    @Test
+    public void testConstructDynamically() {
+        decompileScriptPack("assembled", "tests.TestConstructDynamically", "package tests\n"
+                + "{\n"
+                + "import flash.display.*;\n"
+                + "import flash.utils.getDefinitionByName;\n"
+                + "public class TestConstructDynamically\n"
+                + "{\n"
+                + "public function TestConstructDynamically()\n"
+                + "{\n"
+                + "super();\n"
+                + "}\n"
+                + "public function test() : void\n"
+                + "{\n"
+                + "var _loc1_:* = new (getDefinitionByName(\"flash.display\"+\".\"+\"Sprite\"))();\n"
+                + "_loc1_ = new (getDefinitionByName(\"Object\"))();\n"
+                + "}\n"
+                + "}\n"
+                + "}");
+    }
 }